some restructuring of how git exporter works
Took some steps towards supporting the export of tags, although it's not
complete and it's on the back-burner. The main thing is that this fixes (I
think) the issue where the workspace was getting messed up when we were
checking out various branches. It should now set your workspace back to the
state it was in, with the exception that no changes are staged (if you had
previously staged some of your changes). More investigation needs to be done
around this to see if this still works when you have conflicting staged and
unstaged changes, and if it works in a variety of scenarios that I may not have
tested.