Change node versions we test with on travis
At this time stable (6.2) is failing because of some dependency that apparently
has an issue with the architecture of the system that travis is running the
build on. Not clear at this time, but also not super important to me. We are
standardizing on 4.4.4 for all silvermine projects currently, so we'll test
that as well as 4 and 5 latest.